requirements.txt
ccxt==4.1.13
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
